Move instrumentation into funsor.instrument module
